+++ bitalarm-impl/src/test/java/com/dwarfeng/bitalarm/impl/service/AlarmHistoryMaintainServiceImplTest.java
@@ -49,6 +49,7 @@ public class AlarmHistoryMaintainServiceImplTest {
             AlarmHistory alarmHistory = new AlarmHistory(
                     null,
                     null,
+                    12450,
                     1,
                     "我是报警信息",
                     (byte) 0,