+++ bitalarm-sdk/src/main/java/com/dwarfeng/bitalarm/sdk/bean/entity/FastJsonAlarmHistory.java
@@ -25,6 +25,7 @@ public class FastJsonAlarmHistory implements Bean {
             return new FastJsonAlarmHistory(
                     FastJsonLongIdKey.of(alarmHistory.getKey()),
                     FastJsonLongIdKey.of(alarmHistory.getAlarmSettingKey()),
+                    alarmHistory.getPointId(),
                     alarmHistory.getIndex(),
                     alarmHistory.getAlarmMessage(),
                     alarmHistory.getAlarmType(),
@@ -41,32 +42,36 @@ public class FastJsonAlarmHistory implements Bean {
     @JSONField(name = "alarm_setting_key", ordinal = 2)
     private FastJsonLongIdKey alarmSettingKey;
 
-    @JSONField(name = "index", ordinal = 3)
+    @JSONField(name = "point_id", ordinal = 3)
+    private long pointId;
+
+    @JSONField(name = "index", ordinal = 4)
     private int index;
 
-    @JSONField(name = "alarm_message", ordinal = 4)
+    @JSONField(name = "alarm_message", ordinal = 5)
     private String alarmMessage;
 
-    @JSONField(name = "alarm_type", ordinal = 5)
+    @JSONField(name = "alarm_type", ordinal = 6)
     private byte alarmType;
 
-    @JSONField(name = "start_date", ordinal = 6)
+    @JSONField(name = "start_date", ordinal = 7)
     private Date startDate;
 
-    @JSONField(name = "end_date", ordinal = 7)
+    @JSONField(name = "end_date", ordinal = 8)
     private Date endDate;
 
-    @JSONField(name = "duration", ordinal = 8)
+    @JSONField(name = "duration", ordinal = 9)
     private long duration;
 
     public FastJsonAlarmHistory() {
     }
 
     public FastJsonAlarmHistory(
-            FastJsonLongIdKey key, FastJsonLongIdKey alarmSettingKey, int index, String alarmMessage, byte alarmType, Date startDate,
-            Date endDate, long duration) {
+            FastJsonLongIdKey key, FastJsonLongIdKey alarmSettingKey, long pointId, int index, String alarmMessage,
+            byte alarmType, Date startDate, Date endDate, long duration) {
         this.key = key;
         this.alarmSettingKey = alarmSettingKey;
+        this.pointId = pointId;
         this.index = index;
         this.alarmMessage = alarmMessage;
         this.alarmType = alarmType;
@@ -89,6 +94,14 @@ public class FastJsonAlarmHistory implements Bean {
 
     public void setAlarmSettingKey(FastJsonLongIdKey alarmSettingKey) {
         this.alarmSettingKey = alarmSettingKey;
+    }
+
+    public long getPointId() {
+        return pointId;
+    }
+
+    public void setPointId(long pointId) {
+        this.pointId = pointId;
     }
 
     public int getIndex() {
@@ -144,6 +157,7 @@ public class FastJsonAlarmHistory implements Bean {
         return "FastJsonAlarmHistory{" +
                 "key=" + key +
                 ", alarmSettingKey=" + alarmSettingKey +
+                ", pointId=" + pointId +
                 ", index=" + index +
                 ", alarmMessage='" + alarmMessage + '\'' +
                 ", alarmType=" + alarmType +