+++ bitalarm-stack/src/main/java/com/dwarfeng/bitalarm/stack/bean/entity/AlarmHistory.java
@@ -17,6 +17,7 @@ public class AlarmHistory implements Entity<LongIdKey> {
 
     private LongIdKey key;
     private LongIdKey alarmSettingKey;
+    private long pointId;
     private int index;
     private String alarmMessage;
     private byte alarmType;
@@ -28,10 +29,11 @@ public class AlarmHistory implements Entity<LongIdKey> {
     }
 
     public AlarmHistory(
-            LongIdKey key, LongIdKey alarmSettingKey, int index, String alarmMessage, byte alarmType, Date startDate,
-            Date endDate, long duration) {
+            LongIdKey key, LongIdKey alarmSettingKey, long pointId, int index, String alarmMessage, byte alarmType,
+            Date startDate, Date endDate, long duration) {
         this.key = key;
         this.alarmSettingKey = alarmSettingKey;
+        this.pointId = pointId;
         this.index = index;
         this.alarmMessage = alarmMessage;
         this.alarmType = alarmType;
@@ -56,6 +58,14 @@ public class AlarmHistory implements Entity<LongIdKey> {
 
     public void setAlarmSettingKey(LongIdKey alarmSettingKey) {
         this.alarmSettingKey = alarmSettingKey;
+    }
+
+    public long getPointId() {
+        return pointId;
+    }
+
+    public void setPointId(long pointId) {
+        this.pointId = pointId;
     }
 
     public int getIndex() {
@@ -111,6 +121,7 @@ public class AlarmHistory implements Entity<LongIdKey> {
         return "AlarmHistory{" +
                 "key=" + key +
                 ", alarmSettingKey=" + alarmSettingKey +
+                ", pointId=" + pointId +
                 ", index=" + index +
                 ", alarmMessage='" + alarmMessage + '\'' +
                 ", alarmType=" + alarmType +